p04-chore: 엔티티 빌더 패턴 도입

+++ p04-concurrent-leaderboard/src/main/java/org/example/p04concurrentleaderboard/refund/Refund.java
@@ -21,8 +21,11 @@ public class Refund {
     @Column(nullable = false)
     private Instant createdAt;
 
+    /**
+     * 구매 반영 후 랭킹
+     */
     @Column
-    private Long rank;
+    private Integer rank;
 
     protected Refund() {
     }
